model-deployment/pyproject.toml
[project]
name = "model-deployment"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "azure-ai-ml>=1.28.1",
    "azure-identity>=1.24.0",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "transformers>=4.55.2",
]
